Sidebar reworked: remove setSidebar state, changed structure to Router DOM
