+++ src/components/Sidebar/Sidebar.module.scss
@@ -7,6 +7,8 @@
   border: 0.5px solid white;
   background: App.$primary;
   margin: 2px 0;
+  text-align: center;
+  text-decoration: none;
 }
 
 .sidebar {
