+++ src/components/Sidebar/Sidebar.tsx
@@ -1,13 +1,34 @@
 import { ReactElement } from "react";
-import Events from "./Events";
-import Options from "./Options";
+import { NavLink, Route, Routes } from "react-router-dom";
+import AllEvents from "./AllEvents";
+import MyEvents from "./MyEvents";
+import NewEvent from "./NewEvent";
+import PartEvents from "./PartEvents";
 import styles from "./Sidebar.module.scss";
 
-export const Sidebar = (): ReactElement => {
+export const Sidebar : React.FC = () => {
   return (
     <div className={styles.sidebar}>
-      <Options />
-      <Events />
+      <div className={styles.options}>
+        <NavLink className={({ isActive }) => (isActive ? styles["options__button--active"] : styles.options__button)} to="my">
+          Moje wydarzenia
+        </NavLink>
+        <NavLink className={({ isActive }) => (isActive ? styles["options__button--active"] : styles.options__button)} to="taken">
+          Biorę udział w...
+        </NavLink>
+        <NavLink className={({ isActive }) => (isActive ? styles["options__button--active"] : styles.options__button)} to="upcomming">
+          Nadchodzące wydarzenia
+        </NavLink>
+        <NavLink className={({ isActive }) => (isActive ? styles["options__button--active"] : styles.options__button)} to="add">
+          Dodaj wydarzenie
+        </NavLink>
+      </div>
+      <Routes>
+        <Route path="my" element={<MyEvents />} />
+        <Route path="taken" element={<PartEvents />} />
+        <Route path="upcomming" element={<AllEvents />} />
+        <Route path="add" element={<NewEvent />} />
+      </Routes>
     </div>
   );
 };
